# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcademiQ Logbook API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Entry {
  id: number;
  userId: string;
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  /** Raw activity text entered by the student */
  rawActivity: string;
  /**
   * AI-rewritten professional logbook entry
   * @nullable
   */
  rewrittenEntry: string | null;
  /**
   * SIWES week number
   * @nullable
   */
  week: number | null;
  /**
   * Day of the week (Monday, Tuesday, etc.)
   * @nullable
   */
  dayOfWeek: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateEntryBody {
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  /** Raw activity text */
  rawActivity: string;
  /**
   * AI-rewritten logbook entry
   * @nullable
   */
  rewrittenEntry?: string | null;
  /**
   * SIWES week number
   * @nullable
   */
  week?: number | null;
  /**
   * Day of the week
   * @nullable
   */
  dayOfWeek?: string | null;
}

export interface UpdateEntryBody {
  date?: string;
  rawActivity?: string;
  /** @nullable */
  rewrittenEntry?: string | null;
  /** @nullable */
  week?: number | null;
  /** @nullable */
  dayOfWeek?: string | null;
}

export interface EntryStats {
  totalEntries: number;
  currentStreak: number;
  longestStreak: number;
  totalWeeks: number;
  entriesThisWeek: number;
}

export interface RewriteEntryBody {
  /** Raw activity description to rewrite */
  rawActivity: string;
  /** Date of the activity */
  date?: string;
  /**
   * SIWES week number for context
   * @nullable
   */
  week?: number | null;
}

export interface RewriteEntryResponse {
  /** Professionally rewritten logbook entry */
  rewrittenEntry: string;
}

export interface UserProfile {
  userId: string;
  /** @nullable */
  fullName: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  school: string | null;
  /** @nullable */
  course: string | null;
  /** @nullable */
  siwesCompany: string | null;
  /** @nullable */
  department: string | null;
  /** @nullable */
  siwesDuration: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UpsertProfileBody {
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  school?: string | null;
  /** @nullable */
  course?: string | null;
  /** @nullable */
  siwesCompany?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  siwesDuration?: string | null;
}

export type ListEntriesParams = {
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
  /**
   * Filter by SIWES week number
   */
  week?: number;
  /**
   * Filter by month (YYYY-MM)
   */
  month?: string;
};