# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AcademiQ Logbook API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all entries for the current user
 */
export const ListEntriesQueryParams = zod.object({
  date: zod.coerce.string().optional(),
  week: zod.coerce.number().optional(),
  month: zod.coerce.string().optional(),
});

export const ListEntriesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  rawActivity: zod
    .string()
    .describe("Raw activity text entered by the student"),
  rewrittenEntry: zod
    .string()
    .nullable()
    .describe("AI-rewritten professional logbook entry"),
  week: zod.number().nullable().describe("SIWES week number"),
  dayOfWeek: zod
    .string()
    .nullable()
    .describe("Day of the week (Monday, Tuesday, etc.)"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListEntriesResponse = zod.array(ListEntriesResponseItem);

/**
 * @summary Create a new logbook entry
 */
export const CreateEntryBody = zod.object({
  date: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  rawActivity: zod.string().describe("Raw activity text"),
  rewrittenEntry: zod.string().nullish().describe("AI-rewritten logbook entry"),
  week: zod.number().nullish().describe("SIWES week number"),
  dayOfWeek: zod.string().nullish().describe("Day of the week"),
});

/**
 * @summary Get entry statistics for the current user
 */
export const GetEntryStatsResponse = zod.object({
  totalEntries: zod.number(),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  totalWeeks: zod.number(),
  entriesThisWeek: zod.number(),
});

/**
 * @summary Get the most recent entries (last 7 days)
 */
export const GetRecentEntriesResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  rawActivity: zod
    .string()
    .describe("Raw activity text entered by the student"),
  rewrittenEntry: zod
    .string()
    .nullable()
    .describe("AI-rewritten professional logbook entry"),
  week: zod.number().nullable().describe("SIWES week number"),
  dayOfWeek: zod
    .string()
    .nullable()
    .describe("Day of the week (Monday, Tuesday, etc.)"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetRecentEntriesResponse = zod.array(GetRecentEntriesResponseItem);

/**
 * @summary Get a single entry
 */
export const GetEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetEntryResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  rawActivity: zod
    .string()
    .describe("Raw activity text entered by the student"),
  rewrittenEntry: zod
    .string()
    .nullable()
    .describe("AI-rewritten professional logbook entry"),
  week: zod.number().nullable().describe("SIWES week number"),
  dayOfWeek: zod
    .string()
    .nullable()
    .describe("Day of the week (Monday, Tuesday, etc.)"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update an entry
 */
export const UpdateEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateEntryBody = zod.object({
  date: zod.string().optional(),
  rawActivity: zod.string().optional(),
  rewrittenEntry: zod.string().nullish(),
  week: zod.number().nullish(),
  dayOfWeek: zod.string().nullish(),
});

export const UpdateEntryResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string().describe("ISO date string (YYYY-MM-DD)"),
  rawActivity: zod
    .string()
    .describe("Raw activity text entered by the student"),
  rewrittenEntry: zod
    .string()
    .nullable()
    .describe("AI-rewritten professional logbook entry"),
  week: zod.number().nullable().describe("SIWES week number"),
  dayOfWeek: zod
    .string()
    .nullable()
    .describe("Day of the week (Monday, Tuesday, etc.)"),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete an entry
 */
export const DeleteEntryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary AI rewrite raw activity into professional SIWES logbook entry
 */
export const RewriteEntryBody = zod.object({
  rawActivity: zod.string().describe("Raw activity description to rewrite"),
  date: zod.string().optional().describe("Date of the activity"),
  week: zod.number().nullish().describe("SIWES week number for context"),
});

export const RewriteEntryResponse = zod.object({
  rewrittenEntry: zod
    .string()
    .describe("Professionally rewritten logbook entry"),
});

/**
 * @summary Get the current user's profile
 */
export const GetProfileResponse = zod.object({
  userId: zod.string(),
  fullName: zod.string().nullable(),
  email: zod.string().nullable(),
  school: zod.string().nullable(),
  course: zod.string().nullable(),
  siwesCompany: zod.string().nullable(),
  department: zod.string().nullable(),
  siwesDuration: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Create or update the current user's profile
 */
export const UpsertProfileBody = zod.object({
  fullName: zod.string().nullish(),
  email: zod.string().nullish(),
  school: zod.string().nullish(),
  course: zod.string().nullish(),
  siwesCompany: zod.string().nullish(),
  department: zod.string().nullish(),
  siwesDuration: zod.string().nullish(),
});

export const UpsertProfileResponse = zod.object({
  userId: zod.string(),
  fullName: zod.string().nullable(),
  email: zod.string().nullable(),
  school: zod.string().nullable(),
  course: zod.string().nullable(),
  siwesCompany: zod.string().nullable(),
  department: zod.string().nullable(),
  siwesDuration: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});